process can now be killed on button press
I'm not happy with how the logic works and there is still lots to
clean up but I wanted to commit this as it's still forward progress
and I will end up refactoring most of the application as I lean more
about Rust.

I was pleased to see how Arc<Mutex<T>> worked in practice. I did
make a deadlock the first time I tried but quickly understood where
it was introduced.

This Rust stuff is fun.